menu/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.14",
    "discord-py[voice]>=2.5.2",
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.39",
    "python-dotenv>=1.1.0",
    "pynacl>=1.5.0",
    "youtube-dl>=2021.12.17",
    "yt-dlp>=2025.3.26",
    "openai>=1.68.2",
    "psutil>=7.0.0",
]
